Tweaked empty constants to fix code gen
Changed emptyList, emptyAray and unit to be private nonterminals (so
they don't appear as types in the parse tree, and the rules are pulled
up into rule usage). This fixes the code gen for constantExpression and
constantPattern, which both expose a value that is either a constant
leaf token, or one of these empty types. Not sure this is right (might
be better to introduce a private aux composite to hold all of these
things, and still keep the code gen happy).

On the plus side. Clean compile!